gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
aioredis==2.0.1
prometheus-client==0.19.0
structlog==23.2.0
python-json-logger==2.0.7